Add files via upload
These excel sheets cover all the accessible water that are within the buffers. Water is described as river, sea, or freshwater. Only freshwater is considered for hydration and used by all bats. Sea water and river (which is brackish, and reportedly not often used by the bats) is more likely useful to only two species of bats, the ones that trawl for food